requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.27.1
pydantic==2.6.4
sentence-transformers==2.5.1
numpy==1.26.4
scikit-learn==1.4.2
streamlit==1.31.1
requests==2.31.0
python-dotenv==1.0.1
tenacity==8.2.3